requirements.txt
langchain-groq
streamlit-lottie
requests